Improve efficiency of ServerSessionPool (#951)
* Replace use of ConcurrentPool with direct use of ConcurrentLinkedDeque
* Only prune on ServerSessionPool#get
* Cache ClusterDescription#logicalSessionTimeoutMinutes

JAVA-4629

+++ driver-legacy/src/test/functional/com/mongodb/Fixture.java
@@ -50,7 +50,7 @@ public final class Fixture {
         return mongoClient;
     }
 
-    public static int getServerSessionPoolInUseCount() {
+    public static long getServerSessionPoolInUseCount() {
         return getMongoClient().getServerSessionPool().getInUseCount();
     }
